Add na class for features
In anticipation of the creation of a single table describing all watches
and features, we need a na class (Not Applicable) to indicate features
for which the hardware does not exist on particular watch faces.  For
instance, dory does not have GPS, so it would have na as the status
rather than bad, good, or unknown, which are the existing classes of
feature support.

Signed-off-by: Ed Beroset <beroset@ieee.org>